Revert "Added init application factory"
