+++ api/wsgi.py
@@ -1,0 +1,4 @@
+from app import application
+
+if __name__ == "__main__":
+    application.run()
