--- wsgi.py
@@ -1,6 +1,0 @@
-from api import create_app
-
-app = create_app()
-
-if __name__ == "__main__":
-    app.run()
